requirements.txt
telethon==1.28.5
sqlalchemy==2.0.20
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pillow==10.0.0
psutil==5.9.5
aiohttp>=3.8.5
jieba>=0.42.1
numpy>=1.26.0
matplotlib>=3.7.2
pandas>=2.0.3
emoji==2.8.0
schedule==1.2.0
colorama==0.4.6
tqdm==4.66.1
tabulate==0.9.0